Rearrange scikit-learn interface
Moved scikit-learn support out of LCS, and combined the learn() and
solve() methods into run() with a flag to indicate which behavior is
desired. The code is not functional in its current state; I plan to
create a branch for the scikit-learn support and remove it from the
master on my next commit.