Removed scikit-learn interface
It has become clear that there is no direct mapping from reinforcement
learning algorithms to scikit-learn's API, particularly in the case of
on-line learning -- at least not based on my understanding. I have
created the scikit-learn branch for further investigation into
supporting this interface, and removed the code from the master branch.