LOCKDOWN: v1.0.0
Barring discovery of a major issue, the interface is now frozen until
v1.0.0 is released. After v1.0.0 is released, all future changes will be
made on branches only, and the major version number will be incremented
on any change that breaks backward compatibility. The full test suite
will be run before branches are merged back in.

This commit pertains directly to issue #21.