BitCondition Bitwise Operator Testing
Added unit testing for bitwise operators for BitCondition. There were
two competing interpretations for bitwise operations on BitConditions,
one being that the & of two conditions should match the & of two
BitStrings the original conditions matched, and the other being that,
provided the bits of the original conditions are compatible, the & of
them should be a condition that matches only the BitStrings both of them
match. The second is the more useful way of treating these operators,
despite it leading to a divergence in behavior between BitStrings and
BitConditions. The new tests ensure that this is the interpretation
used.